Move ThermoTable back out of the thermo subfolder.
