+++ src/stanshock/physics/thermotable.py
